pyproject.toml
[tool.black]
preview = true

[tool.poetry]
name = "hikka"
version = "1.7.9"
description = "Telegram userbot"
authors = ["Friendly Telegram, Dan Gazizullin, codrago, kramiikk"]
license = "GNU AGPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
gitpython = "^3.1.44"
meval = "^2.5"
requests = "^2.32.3"
grapheme = "^0.6.0"
emoji = "^2.1.0"
psutil = "^6.1.1"
tgcrypto = "^1.2.5"
hikka-tl-new = "^2.0.8"
rsa = "^4.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
